Projeto Final

+++ lib/components/user_tile.dart
@@ -1,14 +1,14 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:atual_controle_usuario/models/user.dart';
-import 'package:atual_controle_usuario/provider/users.dart';
 import 'package:atual_controle_usuario/routes/app_routes.dart';
-import 'package:provider/provider.dart';
+
+import '../globals.dart';
 
 class UserTile extends StatelessWidget {
   final User user;
 
-  const UserTile(this.user);
+  UserTile(this.user);
 
   @override
   Widget build(BuildContext context) {
@@ -37,33 +37,37 @@ class UserTile extends StatelessWidget {
               icon: Icon(Icons.delete),
               color: Colors.red,
               onPressed: () {
-                showDialog(
-                  context: context,
-                  builder: (ctx) => AlertDialog(
-                    title: Text('Excluir Usuário'),
-                    content: Text('Tem certeza disto?'),
-                    actions: <Widget>[
-                      FlatButton(
-                        child: Text('Não'),
-                        onPressed: () => Navigator.of(context).pop(false),
+                db.collection("users").doc(idUser).get().then((value) {
+                //Vejo se o usuario logado é o mesmo que tá tentando deletar
+                  if (value.data()["email"] != user.email) {
+                    showDialog(
+                      context: context,
+                      builder: (ctx) => AlertDialog(
+                        title: Text('Excluir Usuário'),
+                        content: Text('Tem certeza disto?'),
+                        actions: <Widget>[
+                          FlatButton(
+                            child: Text('Não'),
+                            onPressed: () => Navigator.of(context).pop(false),
+                          ),
+                          FlatButton(
+                            child: Text('Sim'),
+                            onPressed: () => Navigator.of(context).pop(true),
+                          ),
+                        ],
                       ),
-                      FlatButton(
-                        child: Text('Sim'),
-                        onPressed: () => Navigator.of(context).pop(true),
-                      ),
-                    ],
-                  ),
-                ).then((confimed) {
-                  if (confimed) {
-                    Provider.of<Users>(context, listen: false).remove(user);
-                    FirebaseFirestore db = FirebaseFirestore.instance;
-                    db
-                        .collection("users")
-                        .where("name", isEqualTo: user.name)
-                        .get()
-                        .then(
-                          (value) => value.docs.single.reference.delete(),
-                        );
+                    ).then((confimed) {
+                      if (confimed) {
+                        FirebaseFirestore db = FirebaseFirestore.instance;
+                        db
+                            .collection("users")
+                            .where("name", isEqualTo: user.name)
+                            .get()
+                            .then(
+                              (value) => value.docs.single.reference.delete(),
+                            );
+                      }
+                    });
                   }
                 });
               },
